requirements.txt
SQLAlchemy==1.4.26
Flask==2.0.2
PyJWT==2.3.0